Automatically generate handlers from Messengers
