+++ pyro/infer/traceenum_elbo.py
@@ -17,7 +17,7 @@ from pyro.infer.util import Dice, is_validation_enabled
 from pyro.ops import packed
 from pyro.ops.contract import contract_tensor_tree, contract_to_tensor
 from pyro.ops.rings import SampleRing
-from pyro.poutine.enumerate_messenger import EnumerateMessenger
+from pyro.poutine.enum_messenger import EnumMessenger
 from pyro.util import check_traceenum_requirements, ignore_jit_warnings, warn_if_nan
 
 
@@ -327,8 +327,8 @@ class TraceEnum_ELBO(ELBO):
         # Enable parallel enumeration over the vectorized guide and model.
         # The model allocates enumeration dimensions after (to the left of) the guide,
         # accomplished by preserving the _ENUM_ALLOCATOR state after the guide call.
-        guide_enum = EnumerateMessenger(first_available_dim=-1 - self.max_plate_nesting)
-        model_enum = EnumerateMessenger()  # preserve _ENUM_ALLOCATOR state
+        guide_enum = EnumMessenger(first_available_dim=-1 - self.max_plate_nesting)
+        model_enum = EnumMessenger()  # preserve _ENUM_ALLOCATOR state
         guide = guide_enum(guide)
         model = model_enum(model)
