+++ pyro/poutine/block_messenger.py
@@ -63,11 +63,8 @@ def _make_default_hide_fn(hide_all, expose_all, hide, expose, hide_types, expose
 
 class BlockMessenger(Messenger):
     """
-    This Messenger selectively hides Pyro primitive sites from the outside world.
+    This handler selectively hides Pyro primitive sites from the outside world.
     Default behavior: block everything.
-    BlockMessenger has a flexible interface that allows users
-    to specify in several different ways
-    which sites should be hidden or exposed.
 
     A site is hidden if at least one of the following holds:
 
@@ -77,47 +74,38 @@ class BlockMessenger(Messenger):
         3. ``msg["name"] not in expose and msg["type"] not in expose_types``
         4. ``hide``, ``hide_types``, and ``expose_types`` are all ``None``
 
-
     For example, suppose the stochastic function fn has two sample sites "a" and "b".
-    Then any poutine outside of BlockMessenger(fn, hide=["a"])
+    Then any effect outside of ``BlockMessenger(fn, hide=["a"])``
     will not be applied to site "a" and will only see site "b":
 
-    .. doctest::
-       :hide:
+        >>> def fn():
+        ...     a = pyro.sample("a", dist.Normal(0., 1.))
+        ...     return pyro.sample("b", dist.Normal(a, 1.))
+        >>> fn_inner = trace(fn)
+        >>> fn_outer = trace(block(fn_inner, hide=["a"]))
+        >>> trace_inner = fn_inner.get_trace()
+        >>> trace_outer  = fn_outer.get_trace()
+        >>> "a" in trace_inner
+        True
+        >>> "a" in trace_outer
+        False
+        >>> "b" in trace_inner
+        True
+        >>> "b" in trace_outer
+        True
 
-       >>> from pyro.poutine.trace_messenger import TraceMessenger
-
-    >>> def fn():
-    ...     a = pyro.sample("a", dist.Normal(0., 1.))
-    ...     return pyro.sample("b", dist.Normal(a, 1.))
-
-    >>> fn_inner = TraceMessenger()(fn)
-    >>> fn_outer = TraceMessenger()(BlockMessenger(hide=["a"])(TraceMessenger()(fn)))
-    >>> trace_inner = fn_inner.get_trace()
-    >>> trace_outer  = fn_outer.get_trace()
-    >>> "a" in trace_inner
-    True
-    >>> "a" in trace_outer
-    False
-    >>> "b" in trace_inner
-    True
-    >>> "b" in trace_outer
-    True
-
-    See the constructor for details.
-
+    :param fn: a stochastic function (callable containing Pyro primitive calls)
     :param hide_fn: function that takes a site and returns True to hide the site
         or False/None to expose it.  If specified, all other parameters are ignored.
         Only specify one of hide_fn or expose_fn, not both.
     :param expose_fn: function that takes a site and returns True to expose the site
         or False/None to hide it.  If specified, all other parameters are ignored.
         Only specify one of hide_fn or expose_fn, not both.
-    :param bool hide_all: hide all sites
-    :param bool expose_all: expose all sites normally
-    :param list hide: list of site names to hide, rest will be exposed normally
-    :param list expose: list of site names to expose, rest will be hidden
-    :param list hide_types: list of site types to hide, rest will be exposed normally
-    :param list expose_types: list of site types to expose normally, rest will be hidden
+    :param hide: list of site names to hide
+    :param expose: list of site names to be exposed while all others hidden
+    :param hide_types: list of site types to be hidden
+    :param expose_types: list of site types to be exposed while all others hidden
+    :returns: stochastic function decorated with a :class:`~pyro.poutine.block_messenger.BlockMessenger`
     """
 
     def __init__(self, hide_fn=None, expose_fn=None,
