+++ pyro/poutine/broadcast_messenger.py
@@ -4,12 +4,34 @@ from .messenger import Messenger
 
 class BroadcastMessenger(Messenger):
     """
-    `BroadcastMessenger` automatically broadcasts the batch shape of
-    the stochastic function at a sample site when inside a single
-    or nested plate context. The existing `batch_shape` must be
-    broadcastable with the size of the :class:`~pyro.plate`
-    contexts installed in the `cond_indep_stack`.
+    Automatically broadcasts the batch shape of the stochastic function
+    at a sample site when inside a single or nested plate context.
+    The existing `batch_shape` must be broadcastable with the size
+    of the :class:`~pyro.plate` contexts installed in the
+    `cond_indep_stack`.
+
+    Notice how `model_automatic_broadcast` below automates expanding of
+    distribution batch shapes. This makes it easy to modularize a
+    Pyro model as the sub-components are agnostic of the wrapping
+    :class:`~pyro.plate` contexts.
+
+    >>> def model_broadcast_by_hand():
+    ...     with IndepMessenger("batch", 100, dim=-2):
+    ...         with IndepMessenger("components", 3, dim=-1):
+    ...             sample = pyro.sample("sample", dist.Bernoulli(torch.ones(3) * 0.5)
+    ...                                                .expand_by(100))
+    ...             assert sample.shape == torch.Size((100, 3))
+    ...     return sample
+
+    >>> @poutine.broadcast
+    ... def model_automatic_broadcast():
+    ...     with IndepMessenger("batch", 100, dim=-2):
+    ...         with IndepMessenger("components", 3, dim=-1):
+    ...             sample = pyro.sample("sample", dist.Bernoulli(torch.tensor(0.5)))
+    ...             assert sample.shape == torch.Size((100, 3))
+    ...     return sample
     """
+
     @staticmethod
     @ignore_jit_warnings(["Converting a tensor to a Python boolean"])
     def _pyro_sample(msg):
