+++ pyro/poutine/infer_config_messenger.py
@@ -3,7 +3,13 @@ from .messenger import Messenger
 
 class InferConfigMessenger(Messenger):
     """
-    Modifies contents of the infer kwarg at sample sites
+    Given a callable that contains Pyro primitive calls
+    and a callable taking a trace site and returning a dictionary,
+    updates the value of the infer kwarg at a sample site to config_fn(site).
+
+    :param fn: a stochastic function (callable containing Pyro primitive calls)
+    :param config_fn: a callable taking a site and returning an infer dict
+    :returns: stochastic function decorated with :class:`~pyro.poutine.infer_config_messenger.InferConfigMessenger`
     """
     def __init__(self, config_fn):
         """
