+++ pyro/poutine/lift_messenger.py
@@ -9,12 +9,32 @@ from .messenger import Messenger
 
 class LiftMessenger(Messenger):
     """
-    Messenger which "lifts" parameters to random samples.
-    Given a stochastic function with param calls and a prior,
-    creates a stochastic function where all param calls are
-    replaced by sampling from prior.
-
+    Given a stochastic function with param calls and a prior distribution,
+    create a stochastic function where all param calls are replaced by sampling from prior.
     Prior should be a callable or a dict of names to callables.
+
+    Consider the following Pyro program:
+
+        >>> def model(x):
+        ...     s = pyro.param("s", torch.tensor(0.5))
+        ...     z = pyro.sample("z", dist.Normal(x, s))
+        ...     return z ** 2
+        >>> lifted_model = lift(model, prior={"s": dist.Exponential(0.3)})
+
+    ``lift`` makes ``param`` statements behave like ``sample`` statements
+    using the distributions in ``prior``.  In this example, site `s` will now behave
+    as if it was replaced with ``s = pyro.sample("s", dist.Exponential(0.3))``:
+
+        >>> tr = trace(lifted_model).get_trace(0.0)
+        >>> tr.nodes["s"]["type"] == "sample"
+        True
+        >>> tr2 = trace(lifted_model).get_trace(0.0)
+        >>> bool((tr2.nodes["s"]["value"] == tr.nodes["s"]["value"]).all())
+        False
+
+    :param fn: function whose parameters will be lifted to random values
+    :param prior: prior function in the form of a Distribution or a dict of stochastic fns
+    :returns: ``fn`` decorated with a :class:`~pyro.poutine.lift_messenger.LiftMessenger`
     """
 
     def __init__(self, prior):
