+++ pyro/poutine/replay_messenger.py
@@ -3,7 +3,30 @@ from .messenger import Messenger
 
 class ReplayMessenger(Messenger):
     """
-    Messenger for replaying from an existing execution trace.
+    Given a callable that contains Pyro primitive calls,
+    return a callable that runs the original, reusing the values at sites in trace
+    at those sites in the new trace
+
+    Consider the following Pyro program:
+
+        >>> def model(x):
+        ...     s = pyro.param("s", torch.tensor(0.5))
+        ...     z = pyro.sample("z", dist.Normal(x, s))
+        ...     return z ** 2
+
+    ``replay`` makes ``sample`` statements behave as if they had sampled the values
+    at the corresponding sites in the trace:
+
+        >>> old_trace = trace(model).get_trace(1.0)
+        >>> replayed_model = replay(model, trace=old_trace)
+        >>> bool(replayed_model(0.0) == old_trace.nodes["_RETURN"]["value"])
+        True
+
+    :param fn: a stochastic function (callable containing Pyro primitive calls)
+    :param trace: a :class:`~pyro.poutine.Trace` data structure to replay against
+    :param params: dict of names of param sites and constrained values
+        in fn to replay against
+    :returns: a stochastic function decorated with a :class:`~pyro.poutine.replay_messenger.ReplayMessenger`
     """
 
     def __init__(self, trace=None, params=None):
